CR feedback - rename file to ImportAliasRenderer.tsx - EntityInsertPanel check for entityDataType of sample type before attempting to query importAliases (since they don't apply to sources)
