+++ packages/components/src/index.ts
@@ -218,10 +218,7 @@ import { DefaultRenderer } from './internal/renderers/DefaultRenderer';
 import { FileColumnRenderer } from './internal/renderers/FileColumnRenderer';
 import { MultiValueRenderer } from './internal/renderers/MultiValueRenderer';
 import { LabelColorRenderer } from './internal/renderers/LabelColorRenderer';
-import {
-    SampleTypeImportAliasRenderer,
-    SourceTypeImportAliasRenderer,
-} from './internal/renderers/SampleTypeImportAliasRenderer';
+import { SampleTypeImportAliasRenderer, SourceTypeImportAliasRenderer } from './internal/renderers/ImportAliasRenderer';
 import { BulkUpdateForm } from './internal/components/forms/BulkUpdateForm';
 import { LabelOverlay } from './internal/components/forms/LabelOverlay';
 import { resolveDetailFieldValue, resolveRenderer } from './internal/components/forms/renderers';
